experiments/preliminary_experiment/requirements.txt
openai>=1.0.0
psycopg2-binary>=2.9.0
mysql-connector-python>=8.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
pandas>=1.5.0
